+++ core/src/main/java/io/micronaut/core/io/service/SoftServiceLoader.java
@@ -17,7 +17,6 @@ package io.micronaut.core.io.service;
 
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.annotation.Nullable;
-import io.micronaut.core.optim.AotHelper;
 import io.micronaut.core.optim.StaticOptimizations;
 import io.micronaut.core.reflect.ClassUtils;
